Update REPORTER_ASSERT/ERRORF to check format strings.
Previously, REPORTER_ASSERT/ERRORF relied on a helper function named
`reporter_string` which papered over zero-argument and one-argument
messages (where one-argument messages are assumed to ignore printf
formatting rules entirely, and just forward the message as-is).

Replacing this helper with a direct call to `SkStringPrintf` allows
the compiler to check format arguments for correctness, but sacrifices
the one-argument special case. In practice the one-argument special
case was very rarely used, so it's not a significant sacrifice,
and this did uncover several real errors in assertion format strings
(including some cases where the wrong number of arguments was passed).

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/506617
Auto-Submit: John Stiles <johnstiles@google.com>

+++ tests/BackendAllocationTest.cpp
@@ -52,7 +52,7 @@ void test_wrapping(GrDirectContext* dContext,
     sk_sp<ManagedBackendTexture> mbet = create(dContext, mipMapped, renderable);
     if (!mbet) {
         ERRORF(reporter, "Couldn't create backendTexture for grColorType %d renderable %s\n",
-               grColorType,
+               (int)grColorType,
                GrRenderable::kYes == renderable ? "yes" : "no");
         return;
     }
@@ -285,7 +285,8 @@ static void check_base_readbacks(GrDirectContext* dContext,
         GrColorInfo info(colorType, kUnpremul_SkAlphaType, nullptr);
         auto surfaceContext = dContext->priv().makeSC(readView, info);
         if (!surfaceContext) {
-            ERRORF(reporter, "Could not create surface context for colorType: %d\n", colorType);
+            ERRORF(reporter, "Could not create surface context for colorType: %d\n",
+                   (int)colorType);
         }
 
         if (!surfaceContext->readPixels(dContext, actual, {0, 0})) {
@@ -293,7 +294,7 @@ static void check_base_readbacks(GrDirectContext* dContext,
             // arbitrary colorType
 #if 0
             ERRORF(reporter, "Couldn't readback from SurfaceContext for colorType: %d\n",
-                   colorType);
+                   (int)colorType);
 #endif
         } else {
             auto name = SkStringPrintf("%s::readPixels",
